# [performance] fix bad performance in aggregator FixedSliceKey