## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] others

micro benchmark was avaliable: https://github.com/stdpain/starrocks-2/commit/28f1b76f7c8931a91d9f3c3d826553c39a18a252
this PR mainly modify was put `assign key size to slice` to a splitted loop

```
key[i].u.size = slice_sizes[i];
```

touch `slice_sizes` when search hash table may be pollute cpu cache, which will cause a bad performace

```
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
BM_agg_hash_map_8                   81207 ns        81161 ns         8655
BM_agg_hash_origin_map_8           173155 ns       173109 ns         4043
BM_agg_hash_map_64                  83997 ns        83977 ns         8320
BM_agg_hash_origin_map_64          187996 ns       187958 ns         3702
BM_agg_hash_map_512                125765 ns       125739 ns         4828
BM_agg_hash_origin_map_512         124676 ns       124632 ns         5018
BM_agg_hash_map_1024               189101 ns       189081 ns         3384
BM_agg_hash_origin_map_1024        184436 ns       184409 ns         3703
BM_agg_hash_map_2048               270996 ns       270980 ns         2767
BM_agg_hash_origin_map_2048        247880 ns       247871 ns         2580
BM_agg_hash_map_4096               365302 ns       365280 ns         2311
BM_agg_hash_origin_map_4096        311834 ns       311844 ns         2301
BM_agg_hash_map_65535              365454 ns       365434 ns         2235
BM_agg_hash_origin_map_65535       326741 ns       326735 ns         2130
BM_agg_hash_map_655350             385887 ns       385905 ns         2035
BM_agg_hash_origin_map_655350      370408 ns       370399 ns         1993
BM_agg_hash_map_6553500            432022 ns       432021 ns         1474
BM_agg_hash_origin_map_6553500     441429 ns       441432 ns         1512
```


cache miss ratio:
```
agg_hash_origin:
       73,436      cache-references:u                                          
       49,609      cache-misses:u            #   67.554 % of all cache refs    
   14,293,755      cycles:u                                                    
   16,052,022      instructions:u            #    1.12  insn per cycle         
    3,827,236      branches:u                                                  

agg_hash_map:
            74,680      cache-references:u                                          
            47,201      cache-misses:u            #   63.204 % of all cache refs    
        13,647,620      cycles:u                                                    
        16,089,079      instructions:u            #    1.18  insn per cycle         
         3,839,552      branches:u        

```